[fix] menuItem 파일 분리
상수를 constants로 분리
